Added getTicketsNum.py
The main script for phase two. Will get all of the issues and count the amount of server names in their descriptions. If an unknown server was found (a.k.a a name that follows "srv-<>" but not in the servers list), it would print that name as an invalid server name. Using matplotlib, it would plot all of the data onto a bar graph.